PROVIDERS: Add custom hooks
